requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-filter==23.3
python-decouple==3.8
